move pre-processing functionality
- preprocess_kh9 -> preprocessing.preprocess_kh9_mc
- rename command line tool from preprocess_kh9 to preprocess_kh9_mc
- move/edit documentation to reflect changes
- preprocess_pan -> preprocessing.preprocess_pan
- add type hinting to other functions in preprocessing
