+++ doc/spymicmac/modules/preprocessing.rst
@@ -2,4 +2,16 @@ spymicmac.preprocessing
 =================================
 
 .. automodule:: spymicmac.preprocessing
-    :members:
+
+KH-9 Hexagon mapping camera
+---------------------------
+
+.. autofunction:: spymicmac.preprocessing.preprocess_kh9_mc
+
+.. autofunction:: spymicmac.preprocessing.initialize_kh9_mc
+
+
+panoramic cameras
+-----------------
+
+.. autofunction:: spymicmac.preprocessing.preprocess_pan
