--- doc/spymicmac/scripts/preprocess_kh9.rst
@@ -1,7 +1,0 @@
-preprocess_kh9
-=================================
-
-.. argparse::
-   :module: spymicmac.tools.preprocess_kh9
-   :func: _argparser
-   :prog: preprocess_kh9
