+++ doc/spymicmac/scripts/preprocess_kh9_mc.rst
@@ -1,0 +1,7 @@
+preprocess_kh9_mc
+=================================
+
+.. argparse::
+   :module: spymicmac.tools.preprocess_kh9_mc
+   :func: _argparser
+   :prog: preprocess_kh9_mc
