+++ pyproject.toml
@@ -60,7 +60,7 @@ join_hexagon = 'spymicmac.tools.join_hexagon:main'
 mosaic_micmac_tiles = 'spymicmac.tools.mosaic_micmac_tiles:main'
 move_bad_tapas = 'spymicmac.tools.move_bad_tapas:main'
 post_process_micmac = 'spymicmac.tools.post_process_micmac:main'
-preprocess_kh9 = 'spymicmac.tools.preprocess_kh9:main'
+preprocess_kh9_mc = 'spymicmac.tools.preprocess_kh9_mc:main'
 preprocess_pan = 'spymicmac.tools.preprocess_pan:main'
 register_relative = 'spymicmac.tools.register_relative:main'
 remove_crosses = 'spymicmac.tools.remove_crosses:main'
