+++ src/spymicmac/tools/preprocess_pan.py
@@ -1,14 +1,5 @@
-import os
 import argparse
-import shutil
-import tarfile
-import numpy as np
-import pandas as pd
-from glob import glob
-from skimage import io, exposure, filters
-from spymicmac.image import join_hexagon, get_parts_list
-from spymicmac.preprocessing import extract
-from spymicmac.resample import crop_panoramic
+from spymicmac.preprocessing import preprocess_pan
 
 
 def _argparser():
@@ -38,11 +29,8 @@ def _argparser():
                         help='The pre-processing steps to skip.')
     parser.add_argument('--tar_ext', action='store', type=str, default='.tgz',
                         help='Extension for tar files (default: .tgz)')
-    parser.add_argument('-m', '--marker_size', action='store', type=int, default=31,
-                        help='The size of the wagon wheel markers to identify in the image (default: 31 px)')
-    parser.add_argument('-s', '--factor', action='store', type=int, default=None,
-                        help='The number by which to divide the image width and height to scale the image '
-                             '(default: do not scale)')
+    parser.add_argument('-r', '--is_reversed', action='store_true',
+                        help='Order of image parts is reversed (a is the right side of the image). (default: False)')
     parser.add_argument('-o', '--overlap', action='store', type=int, default=None,
                         help='The amount of overlap between image parts to use to search for matches. Default depends'
                              'on flavor: KH4 -> 8000, KH9 -> 1000')
@@ -50,8 +38,11 @@ def _argparser():
                     help='the number of rows each search sub-block should cover [2000].')
     parser.add_argument('-b', '--blend', action='store_true',
                         help='Blend across image halves to prevent a sharp line at edge.')
-    parser.add_argument('-r', '--reversed', action='store_true',
-                        help='Order of image parts is reversed (a is the right side of the image). (default: False)')
+    parser.add_argument('-m', '--marker_size', action='store', type=int, default=31,
+                        help='The size of the wagon wheel markers to identify in the image (default: 31 px)')
+    parser.add_argument('-s', '--factor', action='store', type=int, default=None,
+                        help='The number by which to divide the image width and height to scale the image '
+                             '(default: do not scale)')
     parser.add_argument('--clip_limit', action='store', type=float, default=0.005,
                         help='Clipping limit, for contrast-limited adaptive histogram equalization. (default: 0.005)')
     return parser
@@ -61,107 +52,7 @@ def main():
     parser = _argparser()
     args = parser.parse_args()
 
-    assert args.flavor in ['KH4', 'KH9'], "flavor must be one of [KH4, KH9]"
-    pattern_dict = {'KH4': 'DS', 'KH9': 'D3C'}
-    patt = pattern_dict[args.flavor]
-
-    proc_steps = ['extract', 'join', 'filter', 'crop', 'balance']
-
-    # check what steps we're running
-    if args.steps == 'all':
-        do_step = [True] * len(proc_steps)
-    else:
-        do_step = [step in args.steps for step in proc_steps]
-    do = dict(zip(proc_steps, do_step))
-
-    if args.skip != 'none':
-        for step in args.skip:
-            if step in do.keys():
-                do.update({step: False})
-
-    # now, do all the steps we were asked to do, in order
-    if do['extract']:
-        imlist = extract(args.tar_ext)
-        imlist = [tfile.split(args.tar_ext)[0] for tfile in imlist]
-        if len(imlist) == 0:
-            imlist = list(set([os.path.splitext(fn)[0].split('_')[0] for fn in glob(f'{patt}*.tif')]))
-            imlist.sort()
-
-    elif any([do['join'], do['filter'], do['crop']]):
-        imlist = list(set([os.path.splitext(fn)[0].split('_')[0] for fn in glob(f'{patt}*.tif')]))
-        imlist.sort()
-
-    else:
-        imlist = [os.path.splitext(fn.split('OIS-Reech_')[1])[0] for fn in glob('OIS*.tif')]
-        imlist.sort()
-
-    print(imlist)
-
-    if do['join']:
-        print('Joining scanned image parts.')
-        os.makedirs('parts', exist_ok=True)
-
-        for fn_img in imlist:
-            print(fn_img)
-            parts_list = get_parts_list(fn_img)
-            if len(parts_list) < 2:
-                continue
-
-            join_args = {'overlap': args.overlap,
-                         'block_size': args.block_size,
-                         'blend': args.blend,
-                         'is_reversed': args.reversed}
-
-            if args.flavor == 'KH4' and args.overlap is None:
-                join_args.update({'overlap': 8000})
-            elif args.flavor == 'KH9' and args.overlap is None:
-                join_args.update({'overlap': 1000})
-
-            join_hexagon(fn_img, **join_args)
-
-            for fn in parts_list:
-                shutil.move(f'{fn_img}_{fn}.tif', 'parts')
-
-    if do['filter']:
-        print('Filtering images with a 1-sigma Gaussian Filter')
-        for fn_img in imlist:
-            print(fn_img)
-            img = io.imread(fn_img + '.tif')
-            filt = filters.gaussian(img, sigma=1, preserve_range=True).astype(np.uint8)
-            io.imsave(fn_img + '.tif', filt)
-
-    if do['crop']:
-        print('Rotating and cropping images')
-        os.makedirs('Orig', exist_ok=True)
-
-        img_params = pd.DataFrame(columns=['fn_img', 'left', 'right', 'top', 'bot', 'angle'])
-
-        for fn_img in imlist:
-            print(fn_img)
-            border, angle = crop_panoramic(fn_img + '.tif', args.flavor, marker_size=args.marker_size,
-                                           fact=args.factor, return_vals=True)
-            left, right, top, bot = border
-
-            shutil.move(fn_img + '.tif', 'Orig')
-
-            row = pd.Series()
-            row['fn_img'] = fn_img
-            row['left'] = int(left)
-            row['right'] = int(right)
-            row['top'] = int(top)
-            row['bot'] = int(bot)
-            row['angle'] = angle
-
-            img_params.loc[len(img_params)] = row
-            img_params.to_csv('crop_parameters.csv', index=False)
-
-    if do['balance']:
-        print('Using CLAHE to balance image contrast')
-        for fn_img in imlist:
-            print('OIS-Reech_' + fn_img)
-            img = io.imread('OIS-Reech_' + fn_img + '.tif')
-            img_adj = 255 * exposure.equalize_adapthist(img, clip_limit=args.clip_limit)
-            io.imsave('OIS-Reech_' + fn_img + '.tif', img_adj.astype(np.uint8))
+    preprocess_pan(**args)
 
 
 if __name__ == "__main__":
